Perf: unset EnableContentResponseOnWrite dotnet#22999
moved ETag tests to the ConcurencyTest class
