+++ test/EFCore.Cosmos.FunctionalTests/CosmosConcurrencyTest.cs
@@ -27,7 +27,8 @@ namespace Microsoft.EntityFrameworkCore.Cosmos
                 ctx => ctx.Customers.Add(
                     new Customer
                     {
-                        Id = "1", Name = "CreatedTwice",
+                        Id = "1",
+                        Name = "CreatedTwice",
                     }));
         }
 
@@ -38,7 +39,8 @@ namespace Microsoft.EntityFrameworkCore.Cosmos
                 ctx => ctx.Customers.Add(
                     new Customer
                     {
-                        Id = "2", Name = "Added",
+                        Id = "2",
+                        Name = "Added",
                     }),
                 ctx => ctx.Customers.Single(c => c.Id == "2").Name = "Updated",
                 ctx => ctx.Customers.Remove(ctx.Customers.Single(c => c.Id == "2")));
@@ -51,10 +53,79 @@ namespace Microsoft.EntityFrameworkCore.Cosmos
                 ctx => ctx.Customers.Add(
                     new Customer
                     {
-                        Id = "3", Name = "Added",
+                        Id = "3",
+                        Name = "Added",
                     }),
                 ctx => ctx.Customers.Single(c => c.Id == "3").Name = "Updated",
                 ctx => ctx.Customers.Single(c => c.Id == "3").Name = "Updated");
+        }
+
+        [ConditionalFact]
+        public async Task Etag_will_return_when_content_response_enabled_false()
+        {
+            await using var testDatabase = CosmosTestStore.CreateInitialized(DatabaseName);
+
+            var customer = new Customer
+            {
+                Id = "4",
+                Name = "Theon",
+            };
+
+            await using (var context = new ConcurrencyContext(CreateOptions(testDatabase, enableContentResponseOnWrite: false)))
+            {
+                await context.Database.EnsureCreatedAsync();
+
+                context.Add(customer);
+
+                await context.SaveChangesAsync();
+            }
+
+            await using (var context = new ConcurrencyContext(CreateOptions(testDatabase, enableContentResponseOnWrite: false)))
+            {
+                var customerFromStore = await context.Set<Customer>().SingleAsync();
+
+                Assert.Equal(customer.Id, customerFromStore.Id);
+                Assert.Equal("Theon", customerFromStore.Name);
+                Assert.Equal(customer.ETag, customerFromStore.ETag);
+
+                context.Remove(customerFromStore);
+
+                await context.SaveChangesAsync();
+            }
+        }
+
+        [ConditionalFact]
+        public async Task Etag_will_return_when_content_response_enabled_true()
+        {
+            await using var testDatabase = CosmosTestStore.Create(DatabaseName);
+
+            var customer = new Customer
+            {
+                Id = "3",
+                Name = "Theon",
+            };
+
+            await using (var context = new ConcurrencyContext(CreateOptions(testDatabase, enableContentResponseOnWrite: true)))
+            {
+                await context.Database.EnsureCreatedAsync();
+
+                context.Add(customer);
+
+                await context.SaveChangesAsync();
+            }
+
+            await using (var context = new ConcurrencyContext(CreateOptions(testDatabase, enableContentResponseOnWrite: true)))
+            {
+                var customerFromStore = await context.Set<Customer>().SingleAsync();
+
+                Assert.Equal(customer.Id, customerFromStore.Id);
+                Assert.Equal("Theon", customerFromStore.Name);
+                Assert.Equal(customer.ETag, customerFromStore.ETag);
+
+                context.Remove(customerFromStore);
+
+                await context.SaveChangesAsync();
+            }
         }
 
         /// <summary>
@@ -136,6 +207,19 @@ namespace Microsoft.EntityFrameworkCore.Cosmos
                     });
             }
         }
+
+        private DbContextOptions CreateOptions(CosmosTestStore testDatabase, bool enableContentResponseOnWrite)
+        {
+            var optionsBuilder = new DbContextOptionsBuilder();
+
+            new DbContextOptionsBuilder().UseCosmos(testDatabase.ConnectionString, testDatabase.Name,
+                b => b.ApplyConfiguration().ContentResponseOnWriteEnabled(enabled: enableContentResponseOnWrite));
+
+            return testDatabase.AddProviderOptions(optionsBuilder)
+                .EnableDetailedErrors()
+                .Options;
+        }
+
 
         public class Customer
         {
